move site header to its own component
