+++ client/src/_shared/index.js
@@ -1,2 +1,3 @@
 export { AuthLayout } from './AuthLayout'
 export { LoggedInLayout } from './LoggedInLayout'
+export { Header } from './Header'
